requirements-dev.txt
-e .

django-debug-toolbar

django-dynamic-fixture
fake-factory

mock
pytest
pytest-cov
pytest-django
pytest-splinter
pytest-xdist
pyinotify
